backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-cors>=4.0.1",
    "marshmallow>=4.0.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-flask>=1.3.0",
]
